requirements.txt
streamlit==1.24.1
streamlit-extras
streamlit-chat
pydub
openai
google.generativeai
ffmpeg
ffprobe